Sample call to `always_true`
This is generated in bitcode. We need to automatically build out the
support for checking against types and the registry, but this works as
long as those checks are hand-coded (and correct) in the semantic
analyser.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4051, branch: user/amscanne/extern_objects/16

+++ src/stdlib/base.c
@@ -1,4 +1,4 @@
-int always_true()
+long long always_true()
 {
   return 1;
 }
